android stable

+++ androidApp/src/main/java/com/larkes/hackathonpriceapp/android/screen/main/views/EditPriceSheetView.kt
@@ -135,7 +135,7 @@ fun EditPriceSheetView(
                                 onEvent(
                                     MainEvent.ScannedPriceChanged(
                                         viewState.scannedPrice.copy(
-                                            price = it.toFloat()
+                                            price = it
                                         )
                                     )
                                 )
